Fix pulse update on backend removal
Remove the pulse info from the stash when a backend is removed.
This is to avoid the scenario where an unhealthy backend stored remains
unhealthy after being scaled down and up again.
In this case the pulse update had no effect as the removed backend was
still found in the stash.